worker/requirements.txt
joblib
kafka-python
redis
